requirements.txt
accelerate==1.11.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
inquirerpy==0.3.4
Jinja2==3.1.3
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pfzy==0.3.4
pillow==10.2.0
prompt_toolkit==3.0.52
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.57.1
triton==3.0.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
ViennaRNA==2.7.0
wcwidth==0.2.14
fastparquet==2024.11.0
